SD intergration in progress

+++ Software/arduino/Adafruit/Adafruit_GFX.cpp
@@ -809,7 +809,7 @@ void Adafruit_GFX::drawChar(int16_t x, int16_t y, unsigned char c,
 #if ARDUINO >= 100
 size_t Adafruit_GFX::write(uint8_t c) {
 #else
-size_t Adafruit_GFX::write(uint8_t c) {
+int Adafruit_GFX::write(uint8_t c) {
 #endif
     if(!gfxFont) { // 'Classic' built-in font
